+++ packages/sagas/src/pending/pending.test.js
@@ -1,5 +1,4 @@
-import { delay } from 'redux-saga';
-import { select, put, call, take } from 'redux-saga/effects';
+import { delay, select, put, call, take } from 'redux-saga/effects';
 import cmf from '@talend/react-cmf';
 import { Map } from 'immutable';
 import pendingMaybeNeeded, {
@@ -34,7 +33,7 @@ describe('test pending status', () => {
 		const gen = pendingMaybeNeeded('', 'streams:create');
 		let pendersCollection = new Map();
 
-		expect(gen.next().value).toEqual(call(delay, PENDING_DELAY_TO_SHOW));
+		expect(gen.next().value).toEqual(delay(PENDING_DELAY_TO_SHOW));
 		expect(gen.next().value).toEqual(call(ensurePendersCollectionExists));
 		expect(gen.next().value).toEqual(select(findPenders));