+++ packages/theme/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @talend/bootstrap-theme
 
+## 6.38.6
+
+### Patch Changes
+
+- 86f208189: chore: upgrade dependencies and align @talend scoped packages to latest
+
 ## 6.38.5
 
 ### Patch Changes